Move EventRegistrationTokenTable<T> to runtime
